WIP, converting metadata menu functions
