+++ context/app/static/js/components/searchPage/MetadataMenu/MetadataMenu.jsx
@@ -2,6 +2,7 @@ import React from 'react';
 import { trackEvent } from 'js/helpers/trackers';
 import MenuItem from '@mui/material/MenuItem';
 import Link from '@mui/material/Link';
+import Stack from '@mui/material/Stack';
 
 import useSearchViewStore from 'js/stores/useSearchViewStore';
 import postAndDownloadFile from 'js/helpers/postAndDownloadFile';
@@ -13,6 +14,8 @@ import { NewWorkspaceDialogFromSelections } from 'js/components/workspaces/NewWo
 import { useMetadataMenu } from 'js/components/entity-search/MetadataMenu/hooks';
 
 import { useSelectableTableStore } from 'js/shared-styles/tables/SelectableTableProvider';
+import useSWRMutation from 'swr/mutation';
+import LinearProgress from '@mui/material/LinearProgress';
 import { StyledDropdownMenuButton, StyledInfoIcon } from './style';
 
 async function fetchAndDownload({ urlPath, uuids, closeMenu, analyticsCategory }) {
@@ -39,6 +42,17 @@ function MetadataMenu({ type, analyticsCategory }) {
 
   const menuID = 'metadata-menu';
 
+  const { trigger, isMutating } = useSWRMutation(
+    () => (selectedRows.size > 0 ? [...selectedRows] : allResultsUUIDs),
+    (uuids) =>
+      fetchAndDownload({
+        urlPath: `/metadata/v0/${lcPluralType}.tsv`,
+        uuids,
+        closeMenu,
+        analyticsCategory,
+      }),
+  );
+
   return (
     <>
       <StyledDropdownMenuButton menuID={menuID}>Metadata</StyledDropdownMenuButton>
@@ -51,20 +65,16 @@ function MetadataMenu({ type, analyticsCategory }) {
             <StyledInfoIcon color="primary" />
           </SecondaryBackgroundTooltip>
         </MenuItem>
-        <MenuItem
-          onClick={() =>
-            fetchAndDownload({
-              urlPath: `/metadata/v0/${lcPluralType}.tsv`,
-              uuids: selectedRows.size > 0 ? [...selectedRows] : allResultsUUIDs,
-              closeMenu,
-              analyticsCategory,
-            })
-          }
-        >
-          Download
-          <SecondaryBackgroundTooltip title="Download a TSV of the table metadata." placement="bottom-start">
-            <StyledInfoIcon color="primary" />
-          </SecondaryBackgroundTooltip>
+        <MenuItem onClick={trigger}>
+          <Stack direction="column" width="100%">
+            <span>
+              Download
+              <SecondaryBackgroundTooltip title="Download a TSV of the table metadata." placement="bottom-start">
+                <StyledInfoIcon color="primary" />
+              </SecondaryBackgroundTooltip>
+            </span>
+            {isMutating && <LinearProgress />}
+          </Stack>
         </MenuItem>
         {isWorkspacesUser && <NewWorkspaceDialogFromSelections />}
       </DropdownMenu>
